Move xcm to primitives
(cherry picked from commit 55eb01b533dea6b909794e0c0955f7b63423e975)
