+++ frame/ethereum/Cargo.toml
@@ -36,6 +36,7 @@ fp-evm = { version = "3.0.0-dev", path = "../../primitives/evm", default-feature
 fp-rpc = { version = "3.0.0-dev", path = "../../primitives/rpc", default-features = false }
 fp-self-contained = { version = "1.0.0-dev", path = "../../primitives/self-contained", default-features = false }
 fp-storage = { version = "2.0.0", path = "../../primitives/storage", default-features = false }
+fp-xcm = { version = "1.0.0-dev", path = "../../primitives/xcm", default-features = false }
 pallet-evm = { version = "6.0.0-dev", path = "../evm", default-features = false }
 
 [dev-dependencies]
@@ -72,6 +73,7 @@ std = [
 	"fp-rpc/std",
 	"fp-self-contained/std",
 	"fp-storage/std",
+	"fp-xcm/std",
 	"pallet-evm/std",
 ]
 runtime-benchmarks = [
